+++ mx.truffleruby/suite.py
@@ -7,7 +7,7 @@ suite = {
             {
                 "name": "regex",
                 "subdir": True,
-                "version": "d24695e267dbff610ec262c5c990560af88bd804",
+                "version": "5bb740508beb175c209804192c87d6f941a2b112",
                 "urls": [
                     {"url": "https://github.com/oracle/graal.git", "kind": "git"},
                     {"url": "https://curio.ssw.jku.at/nexus/content/repositories/snapshots", "kind": "binary"},
@@ -16,7 +16,7 @@ suite = {
             {
                 "name": "sulong",
                 "subdir": True,
-                "version": "d24695e267dbff610ec262c5c990560af88bd804",
+                "version": "5bb740508beb175c209804192c87d6f941a2b112",
                 "urls": [
                     {"url": "https://github.com/oracle/graal.git", "kind": "git"},
                     {"url": "https://curio.ssw.jku.at/nexus/content/repositories/snapshots", "kind": "binary"},