+++ src/main/java/org/truffleruby/language/RubyCheckArityRootNode.java
@@ -11,6 +11,7 @@ package org.truffleruby.language;
 
 import com.oracle.truffle.api.CompilerDirectives;
 import com.oracle.truffle.api.CompilerDirectives.CompilationFinal;
+import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
 import com.oracle.truffle.api.nodes.Node;
 import org.truffleruby.RubyContext;
 import org.truffleruby.RubyLanguage;
@@ -56,39 +57,29 @@ public abstract class RubyCheckArityRootNode extends RubyRootNode {
     protected void checkArity(VirtualFrame frame) {
         int given = RubyArguments.getPositionalArgumentsCount(frame, keywordArguments);
 
-        if (!keywordArguments) {
-            if (!arityForCheck.check(given)) {
-                if (!checkArityProfile) {
-                    CompilerDirectives.transferToInterpreterAndInvalidate();
-                    checkArityProfile = true;
-                }
-
-                checkArityError(arityForCheck, given, this);
-            }
-        } else {
-            if (!arityForCheck.basicCheck(given)) {
-                if (!checkArityProfile) {
-                    CompilerDirectives.transferToInterpreterAndInvalidate();
-                    checkArityProfile = true;
-                }
-
-                checkArityError(arityForCheck, given, this);
+        if (!arityForCheck.checkPositionalArguments(given)) {
+            if (!checkArityProfile) {
+                CompilerDirectives.transferToInterpreterAndInvalidate();
+                checkArityProfile = true;
             }
 
-            if (checkKeywordArityNode != null) {
-                checkKeywordArityNode.checkArity(frame);
-            }
+            throw checkArityError(arityForCheck, given, this);
+        }
+
+        if (checkKeywordArityNode != null) {
+            checkKeywordArityNode.checkArity(frame);
         }
     }
 
-    public static void checkArityError(Arity arity, int given, Node currentNode) {
+    @TruffleBoundary
+    public static RaiseException checkArityError(Arity arity, int given, Node currentNode) {
         final RubyContext context = RubyContext.get(currentNode);
         if (arity.hasRest()) {
-            throw new RaiseException(
+            return new RaiseException(
                     context,
                     context.getCoreExceptions().argumentErrorPlus(given, arity.getRequired(), currentNode));
         } else if (arity.getOptional() > 0) {
-            throw new RaiseException(
+            return new RaiseException(
                     context,
                     context.getCoreExceptions().argumentError(
                             given,
@@ -96,7 +87,7 @@ public abstract class RubyCheckArityRootNode extends RubyRootNode {
                             arity.getOptional(),
                             currentNode));
         } else {
-            throw new RaiseException(
+            return new RaiseException(
                     context,
                     context.getCoreExceptions().argumentError(given, arity.getRequired(), currentNode));
         }