+++ src/main/java/org/truffleruby/language/dispatch/LiteralCallNode.java
@@ -10,6 +10,7 @@
 package org.truffleruby.language.dispatch;
 
 import com.oracle.truffle.api.CompilerDirectives;
+import com.oracle.truffle.api.HostCompilerDirectives.InliningCutoff;
 import org.truffleruby.core.array.ArrayUtils;
 import org.truffleruby.core.hash.HashNodes.CopyHashAndSetRuby2KeywordsNode;
 import org.truffleruby.core.hash.RubyHash;
@@ -100,14 +101,19 @@ public abstract class LiteralCallNode extends RubyContextSourceNode {
     // NOTE: args is either frame args or user args
     public void copyRuby2KeywordsHash(Object[] args, SharedMethodInfo info) {
         if (!info.getArity().hasRest()) { // https://bugs.ruby-lang.org/issues/18625
-            if (copyHashAndSetRuby2KeywordsNode == null) {
-                CompilerDirectives.transferToInterpreterAndInvalidate();
-                copyHashAndSetRuby2KeywordsNode = insert(CopyHashAndSetRuby2KeywordsNode.create());
-            }
-
-            final RubyHash lastArgument = (RubyHash) ArrayUtils.getLast(args);
-            ArrayUtils.setLast(args, copyHashAndSetRuby2KeywordsNode.execute(lastArgument, false));
+            copyRuby2KeywordsHashBoundary(args);
         }
+    }
+
+    @InliningCutoff
+    private void copyRuby2KeywordsHashBoundary(Object[] args) {
+        if (copyHashAndSetRuby2KeywordsNode == null) {
+            CompilerDirectives.transferToInterpreterAndInvalidate();
+            copyHashAndSetRuby2KeywordsNode = insert(CopyHashAndSetRuby2KeywordsNode.create());
+        }
+
+        final RubyHash lastArgument = (RubyHash) ArrayUtils.getLast(args);
+        ArrayUtils.setLast(args, copyHashAndSetRuby2KeywordsNode.execute(lastArgument, false));
     }
 
 }