+++ src/main/java/org/truffleruby/language/objects/LogicalClassNode.java
@@ -9,7 +9,9 @@
  */
 package org.truffleruby.language.objects;
 
+import com.oracle.truffle.api.HostCompilerDirectives.InliningCutoff;
 import com.oracle.truffle.api.dsl.Cached;
+import com.oracle.truffle.api.dsl.TypeSystemReference;
 import org.truffleruby.core.encoding.RubyEncoding;
 import org.truffleruby.core.klass.RubyClass;
 import org.truffleruby.core.numeric.RubyBignum;
@@ -18,6 +20,7 @@ import org.truffleruby.core.regexp.RubyRegexp;
 import org.truffleruby.core.symbol.RubySymbol;
 import org.truffleruby.core.string.ImmutableRubyString;
 import org.truffleruby.language.Nil;
+import org.truffleruby.language.NoImplicitCastsToLong;
 import org.truffleruby.language.RubyBaseNode;
 import org.truffleruby.language.RubyDynamicObject;
 
@@ -25,6 +28,7 @@ import com.oracle.truffle.api.dsl.GenerateUncached;
 import com.oracle.truffle.api.dsl.Specialization;
 
 @GenerateUncached
+@TypeSystemReference(NoImplicitCastsToLong.class)
 public abstract class LogicalClassNode extends RubyBaseNode {
 
     public static LogicalClassNode create() {
@@ -102,6 +106,7 @@ public abstract class LogicalClassNode extends RubyBaseNode {
         return object.getLogicalClass();
     }
 
+    @InliningCutoff
     @Specialization(guards = "isForeignObject(object)")
     protected RubyClass logicalClassForeign(Object object,
             @Cached ForeignClassNode foreignClassNode) {